Move MiniView Into Separate File
