+++ frontend/src/components/dashboard/miniview.css
@@ -1,11 +1,4 @@
 @import url("../../main.css");
-.dashboard {
-  padding: 3vh 1vw;
-  display: flex;
-  gap: 1.5%;
-  align-items: flex-start;
-  background-color: var(--color-background2);
-}
 
 .dashboard .mini-view {
   border-radius: 4px;
